improve module structure, cleanups, renames
