+++ iroh-willow/src/lib.rs
@@ -3,6 +3,7 @@
 #![allow(missing_docs)]
 
 pub mod actor;
+pub mod auth;
 pub mod engine;
 pub mod form;
 pub mod net;
