+++ iroh-willow/src/session/error.rs
@@ -15,7 +15,7 @@ pub enum Error {
     #[error("local store failed: {0}")]
     Store(#[from] anyhow::Error),
     #[error("authentication error: {0}")]
-    Auth(#[from] crate::store::auth::AuthError),
+    Auth(#[from] crate::auth::AuthError),
     #[error("payload store failed: {0}")]
     PayloadStore(std::io::Error),
     #[error("payload digest does not match expected digest")]
